Remove coach-based agent
The coach framework is probably great if you want to fiddle with
hyperparameters, try different agent-exploration strategy combinations
etc. It didn't work well for me however, since it makes a lot of
assumptions I am violating (the main one being that there is a constant
amount of actions). As a result, the framework works against me instead
of for me.