Make MLP shapes configurable
Requires vendoring the EncodeProcessDecode model, as the one I was using
previously was only intended as a demo and hardcodes these variables.

This is still exposing a lot less flexibility than would be possible,
since each MLP is assumed to have the same shape. In theory, every MLP
could have an individual shape.